[Finish]

+++ bookers2/app/controllers/books_controller.rb
@@ -6,32 +6,60 @@ class BooksController < ApplicationController
   def create
     @book = Book.new(book_params)
     @book.user_id = current_user.id
-    @book.save
-    redirect_to books_path
+    if @book.save
+      flash[:notice] = "You have created book successfully."
+      redirect_to book_path(@book.id)
+    else
+      @user = current_user
+      @books = Book.all
+      render :index
+    end
   end
 
   def index
+    @user = current_user
+    @book = Book.new
     @books = Book.all
   end
 
   def show
     @book = Book.find(params[:id])
+    @new_book = Book.new
+    @user = @book.user
   end
 
   def edit
+    book = Book.find(params[:id])
+    unless current_user == book.user
+      redirect_to books_path
+    end
     @book = Book.find(params[:id])
   end
-  
+
   def update
+    book = Book.find(params[:id])
+    unless current_user == book.user
+      redirect_to books_path
+    end
+
     @book = Book.find(params[:id])
-    @book.update(book_params)
-    redirect_to book_path(@book.id)
+    if @book.update(book_params)
+      flash[:notice] = "You have updated book successfully."
+      redirect_to book_path(@book.id)
+    else
+      render :edit
+    end
   end
 
   def destroy
     book = Book.find(params[:id])
+    unless current_user == book.user
+      redirect_to books_path
+    end
+
+    book = Book.find(params[:id])
     book.destroy
-    redirect_to root_path
+    redirect_to books_path
   end
 
   private
